refactor(googlereader): move constants to separate files
